+++ Scripts/Editor/TMP_SpriteAssetMenu.cs
@@ -14,8 +14,14 @@ namespace TMPro.EditorUtilities
     public static class TMP_SpriteAssetMenu
     {
         // Add a Context Menu to the Sprite Asset Editor Panel to Create and Add a Default Material.
+        [MenuItem("CONTEXT/TMP_SpriteAsset/Add Default Material", true, 2200)]
+        static bool AddDefaultMaterialValidate(MenuCommand command)
+        {
+            return AssetDatabase.IsOpenForEdit(command.context);
+        }
+
         [MenuItem("CONTEXT/TMP_SpriteAsset/Add Default Material", false, 2200)]
-        static void CopyTexture(MenuCommand command)
+        static void AddDefaultMaterial(MenuCommand command)
         {
             TMP_SpriteAsset spriteAsset = (TMP_SpriteAsset)command.context;
 
@@ -28,6 +34,12 @@ namespace TMPro.EditorUtilities
         }
 
         // Add a Context Menu to the Sprite Asset Editor Panel to update existing sprite assets.
+        [MenuItem("CONTEXT/TMP_SpriteAsset/Update Sprite Asset", true, 2100)]
+        static bool UpdateSpriteAssetValidate(MenuCommand command)
+        {
+            return AssetDatabase.IsOpenForEdit(command.context);
+        }
+
         [MenuItem("CONTEXT/TMP_SpriteAsset/Update Sprite Asset", false, 2100)]
         static void UpdateSpriteAsset(MenuCommand command)
         {