+++ Scripts/Runtime/TMP_DefaultControls.cs
@@ -187,7 +187,8 @@ namespace TMPro
             SetDefaultColorTransitionValues(inputField);
 
             // Use UI.Mask for Unity 5.0 - 5.1 and 2D RectMask for Unity 5.2 and up
-            textArea.AddComponent<RectMask2D>();
+            RectMask2D rectMask = textArea.AddComponent<RectMask2D>();
+            rectMask.padding = new Vector4(-8, -5, -8, -5);
 
             RectTransform textAreaRectTransform = textArea.GetComponent<RectTransform>();
             textAreaRectTransform.anchorMin = Vector2.zero;