+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 These are the release notes for the TextMesh Pro UPM package which was first introduced with Unity 2018.1. Please see the following link for the Release Notes for prior versions of TextMesh Pro. http://digitalnativestudios.com/forum/index.php?topic=1363.0
 
-## [2.1.0-preview.3] - 2019-12-16
+## [3.0.0-preview.3] - 2019-12-16
+## [2.1.0-preview.3]
 ## [1.5.0-preview.3]
 ### Changes
 - Fixed potential issue with TMP Dropdown where calling Show() and Hide() in very short interval could result in additional Blockers. Case #1194114
@@ -21,6 +22,8 @@ These are the release notes for the TextMesh Pro UPM package which was first int
 - Revised the Create Material Preset context menu option to issue a warning and ignore materials outside the project. Case #1200109
 - Added experimental ITextPreprocessor interface to allow users to create custom components to handle text preprocessing and shaping. This interface includes a PreprocessText(string text) function that is called when the object contains a component that inherits from this interface.   
 - Added support for Unity Presets in the Editor for both <TextMeshPro> and <TextMeshProUGUI> components. Case #1191793
+- Fixed missing CanvasRenderer component issue on the Input Field Caret object.
+- Added padding to the 2DRectMask on the TMP Input Field - Text Area object.
 - Optimization to ensure the TMP Update Manager only rebuilds text objects once per frame regardless of the number of cameras in the scene.
 
 ## [2.1.0-preview.2] - 2019-10-30